Aula - Implementando outras exceções - Parte 3

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,13 +3,16 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DataBaseException;
 import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
 
 @Service
 public class ProductService {
@@ -54,26 +57,36 @@ public class ProductService {
         return new ProductDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("Falha de integridade referencial");
+        }
     }
 
     @Transactional
     public ProductDTO update(Long id , ProductDTO dto){
 
-        /* getReferenceById não busca os dados no banco de dados, ele prepara os
-         * dados para atualizar, criando uma instância do objeto com o id e depois o dto atualiza a entity
-         * atráves do método copyDtoToEntity
-         * */
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto,entity);
+        try {
+            /* getReferenceById não busca os dados no banco de dados, ele prepara os
+             * dados para atualizar, criando uma instância do objeto com o id e depois o dto atualiza a entity
+             * atráves do método copyDtoToEntity
+             * */
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto,entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException E){
+                throw new ResourceNotFoundException("Recurso não encontrado");
+        }
 
-        entity = repository.save(entity);
-
-        return new ProductDTO(entity);
     }
-
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
